Do not define a data format, only conventions for localization
There are multiple data formats already in use (PO files, .properties files,
INI files, CSV files, JSON files...) with various levels of integration
depending on the programming language. The goal of this project is not
to mandate a new and different data format, which would require to
convert all existing translations to a new format to start using the API.

My intent is to piggyback on existing projects, defining a set of
conventions and an API to simplify i18n and L10n in complex use cases
(multiple parameters to replace, plurals, gender...).

I also added a new section for REFERENCES, which will collect links
to separate specifications.