Use explicit empty strings to void blocks, use null as default value
The blocks that make parts of templates can now be silenced in a simpler
way by having the application provide an empty string for the key. This
value takes precedence over the template value, which is only looked up
when no value (null) is provided by the application for the key.

This simplifies the complex example greatly, which is shortened by the
fact that there is no need to parse a template that will eventually be
skipped.

The null value is now assigned the role of the default value which allows
to keep a parameter unreplaced, when no matching message is found either.